# Final_Project-OpenAPI.yaml
openapi: 3.0.0
info:
  title: FINAL PROJECT
  contact: {}
  version: "1.0"
servers:
  - url: http://localhost:8000/v1/api
    variables: {}
  - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api
    variables: {}
paths:
...
  /tickets/{id}:
    get:
      tags:
        - TICKET FOR USERS
      summary: GET TICKET BY ID
      operationId: GETTICKETBYID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
...
  /wishlists/{id}:
    get:
      tags:
        - WISHLIST FOR USERS
      summary: GET WISHLIST BY ID
      operationId: GETWISHLISTBYID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
  /wishlists/create:
    post:
      tags:
        - WISHLIST FOR USERS
      summary: CREATE WISHLIST
      operationId: CREATEWISHLIST
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CREATEWISHLISTREQUEST"
                - example:
                    ticket_id_departure: 1
                    ticket_id_return: 2
            example:
              ticket_id_departure: 1
              ticket_id_return: 2
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
  /wishlists/delete/{id}:
    delete:
      tags:
        - WISHLIST FOR USERS
      summary: DELETE WISHLIST
      operationId: DELETEWISHLIST
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
...
  /passanger/delete/{id}:
    delete:
      tags:
        - CRUD PASSANGER
      summary: DELETE PASSANGER
      operationId: DELETEPASSANGER
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
...
  /wallet/create:
    post:
      tags:
        - CRUD WALLET
      summary: CREATE WALLET
      operationId: CREATEWALLET
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CREATEWALLETREQUEST"
                - example:
                    user_id: 3
                    balance: 6000000
            example:
              user_id: 3
              balance: 6000000
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
    delete:
      tags:
        - CRUD WALLET
      summary: DELETE WALLET
      operationId: DeleteDELETEWALLET
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
  /wallet/edit/{id}:
    put:
      tags:
        - CRUD WALLET
      summary: UPDATE WALLET
      operationId: UPDATEWALLET
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UPDATEWALLETREQUEST"
                - example:
                    balance: 10000
            example:
              balance: 10000
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
  /wallet/delete/{id}:
    delete:
      tags:
        - CRUD WALLET
      summary: DELETE WALLET
      operationId: DELETEWALLET
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
...
components:
  schemas:
...
    CREATEWISHLISTREQUEST:
      title: CREATEWISHLISTREQUEST
      required:
        - ticket_id_departure
        - ticket_id_return
      type: object
      properties:
        ticket_id_departure:
          type: integer
          format: int32
        ticket_id_return:
          type: integer
          format: int32
      example:
        ticket_id_departure: 1
        ticket_id_return: 2
...
    CREATEWALLETREQUEST:
      title: CREATEWALLETREQUEST
      required:
        - user_id
        - balance
      type: object
      properties:
        user_id:
          type: integer
          format: int32
        balance:
          type: integer
          format: int32
      example:
        user_id: 3
        balance: 6000000
    UPDATEWALLETREQUEST:
      title: UPDATEWALLETREQUEST
      required:
        - balance
      type: object
      properties:
        balance:
          type: integer
          format: int32
      example:
        balance: 10000
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...